chore(simple-list): renamed pf-listbox to match upstream patternfly simple-list
