+++ elements/package.json
@@ -36,9 +36,9 @@
     "./pf-jump-links/pf-jump-links.js": "./pf-jump-links/pf-jump-links.js",
     "./pf-label/BaseLabel.js": "./pf-label/BaseLabel.js",
     "./pf-label/pf-label.js": "./pf-label/pf-label.js",
-    "./pf-listbox/pf-listbox.js": "./pf-listbox/pf-listbox.js",
-    "./pf-listbox/pf-listbox-group.js": "./pf-listbox/pf-listbox-group.js",
-    "./pf-listbox/pf-listbox-option.js": "./pf-listbox/pf-listbox-option.js",
+    "./pf-simple-list/pf-simple-list.js": "./pf-simple-list/pf-simple-list.js",
+    "./pf-simple-list/pf-simple-list-group.js": "./pf-simple-list/pf-simple-list-group.js",
+    "./pf-simple-list/pf-simple-list-option.js": "./pf-simple-list/pf-simple-list-option.js",
     "./pf-modal/pf-modal.js": "./pf-modal/pf-modal.js",
     "./pf-panel/pf-panel.js": "./pf-panel/pf-panel.js",
     "./pf-progress-stepper/pf-progress-step.js": "./pf-progress-stepper/pf-progress-step.js",
