--- elements/pf-listbox/README.md
@@ -1,11 +1,0 @@
-# Listbox
-Add a description of the component here.
-
-## Usage
-Describe how best to use this web component along with best practices.
-
-```html
-<pf-listbox>
-
-</pf-listbox>
-```
