--- elements/pf-listbox/demo/filter-modes.html
@@ -1,43 +1,0 @@
-<link rel="stylesheet" href="demo.css">
-<script type="module" src="pf-listbox.js"></script>
-
-<h2>filter-mode="disabled"</h2>
-<p>
-  Any arrow keys work. 
-  Only one option can be selected. 
-  Filter is disabled.</p>
-<pf-listbox filter-mode="disabled">
-  <pf-listbox-option>Blue</pf-listbox-option>
-  <pf-listbox-option>Green</pf-listbox-option>
-  <pf-listbox-option>Magenta</pf-listbox-option>
-  <pf-listbox-option>Orange</pf-listbox-option>
-  <pf-listbox-option>Purple</pf-listbox-option>
-  <pf-listbox-option>Pink</pf-listbox-option>
-  <pf-listbox-option>Red</pf-listbox-option>
-  <pf-listbox-option>Yellow</pf-listbox-option>
-</pf-listbox>
-
-<h2>filter-mode="required": combobox example</h2>
-<p>
-  Any arrow keys work. 
-  Only one option can be selected. 
-  Listbox wont appear until input has text.
-</p>
-<label id="combo">
-  Pick a color:
-  <input id="text"
-    aria-controls="listbox"
-    aria-role="combobox" 
-    aria-autocomplete="both" 
-    aria-expanded="false"/>
-  <pf-listbox id="listbox" filter-mode="required">
-    <pf-listbox-option>Blue</pf-listbox-option>
-    <pf-listbox-option>Green</pf-listbox-option>
-    <pf-listbox-option>Magenta</pf-listbox-option>
-    <pf-listbox-option>Orange</pf-listbox-option>
-    <pf-listbox-option>Purple</pf-listbox-option>
-    <pf-listbox-option>Pink</pf-listbox-option>
-    <pf-listbox-option>Red</pf-listbox-option>
-    <pf-listbox-option>Yellow</pf-listbox-option>
-  </pf-listbox>
-</label>
