--- elements/pf-listbox/demo/grouped-options.html
@@ -1,23 +1,0 @@
-<link rel="stylesheet" href="demo.css">
-<script type="module" src="pf-listbox.js"></script>
-
-<p>
-  Any arrow keys work. 
-  Only one option can be selected. 
-  Filter by typing a letter. 
-  Headings are not selectable.
-</p>
-<pf-listbox>
-  <pf-listbox-group>
-    <span slot="group-heading">Breakfast</span>
-    <pf-listbox-option>Eggs</pf-listbox-option>
-    <pf-listbox-option>Toast</pf-listbox-option>
-    <pf-listbox-option>Waffles</pf-listbox-option>
-  </pf-listbox-group>
-  <pf-listbox-group>
-    <span slot="group-heading">Lunch</span>
-    <pf-listbox-option>Salad</pf-listbox-option>
-    <pf-listbox-option>Sandwich</pf-listbox-option>
-    <pf-listbox-option>Soup</pf-listbox-option>
-  </pf-listbox-group>
-</pf-listbox>
