--- elements/pf-listbox/demo/multiple-selections.html
@@ -1,19 +1,0 @@
-<link rel="stylesheet" href="demo.css">
-<script type="module" src="pf-listbox.js"></script>
-
-<p>
-  Multiple options can be selected. Any arrow keys work. 
-  <kbd>Shift</kbd> will toggling off multiple items. 
-  <kbd>Ctrl+A</kbd> will toggle selection on all items. 
-  Filter by typing a letter.
-</p>
-<pf-listbox multiple>
-  <pf-listbox-option>Blue</pf-listbox-option>
-  <pf-listbox-option>Green</pf-listbox-option>
-  <pf-listbox-option>Magenta</pf-listbox-option>
-  <pf-listbox-option>Orange</pf-listbox-option>
-  <pf-listbox-option>Purple</pf-listbox-option>
-  <pf-listbox-option>Pink</pf-listbox-option>
-  <pf-listbox-option>Red</pf-listbox-option>
-  <pf-listbox-option>Yellow</pf-listbox-option>
-</pf-listbox>
