--- elements/pf-listbox/demo/orientation.html
@@ -1,28 +1,0 @@
-<link rel="stylesheet" href="demo.css">
-<script type="module" src="pf-listbox.js"></script>
-
-<h2>orientation="vertical"</h2>
-<p>Only vertical arrow keys work. Filter by typing a letter.</p>
-<pf-listbox orientation="vertical">
-  <pf-listbox-option>Blue</pf-listbox-option>
-  <pf-listbox-option>Green</pf-listbox-option>
-  <pf-listbox-option>Magenta</pf-listbox-option>
-  <pf-listbox-option>Orange</pf-listbox-option>
-  <pf-listbox-option>Purple</pf-listbox-option>
-  <pf-listbox-option>Pink</pf-listbox-option>
-  <pf-listbox-option>Red</pf-listbox-option>
-  <pf-listbox-option>Yellow</pf-listbox-option>
-</pf-listbox>
-
-<h2>orientation="horizontal"</h2>
-<p>Only horizontal arrow keys work. Filter by typing a letter.</p>
-<pf-listbox orientation="horizontal">
-  <pf-listbox-option>Blue</pf-listbox-option>
-  <pf-listbox-option>Green</pf-listbox-option>
-  <pf-listbox-option>Magenta</pf-listbox-option>
-  <pf-listbox-option>Orange</pf-listbox-option>
-  <pf-listbox-option>Purple</pf-listbox-option>
-  <pf-listbox-option>Pink</pf-listbox-option>
-  <pf-listbox-option>Red</pf-listbox-option>
-  <pf-listbox-option>Yellow</pf-listbox-option>
-</pf-listbox>
