--- elements/pf-listbox/demo/pf-listbox.html
@@ -1,19 +1,0 @@
-<link rel="stylesheet" href="demo.css">
-<script type="module" src="pf-listbox.js"></script>
-
-<h2>default</h2>
-<p>
-  Any arrow keys work. 
-  Only one option can be selected. 
-  Filter by typing a letter.
-</p>
-<pf-listbox>
-  <pf-listbox-option>Blue</pf-listbox-option>
-  <pf-listbox-option>Green</pf-listbox-option>
-  <pf-listbox-option>Magenta</pf-listbox-option>
-  <pf-listbox-option>Orange</pf-listbox-option>
-  <pf-listbox-option>Purple</pf-listbox-option>
-  <pf-listbox-option>Pink</pf-listbox-option>
-  <pf-listbox-option>Red</pf-listbox-option>
-  <pf-listbox-option>Yellow</pf-listbox-option>
-</pf-listbox>
