+++ elements/pf-simple-list/demo/demo.css
@@ -1,4 +1,4 @@
-[data-demo="pf-listbox"] {
+[data-demo="pf-simple-list"] {
   padding: 20px;
 }
