+++ elements/pf-simple-list/demo/grouped-options.html
@@ -1,0 +1,23 @@
+<link rel="stylesheet" href="demo.css">
+<script type="module" src="pf-simple-list.js"></script>
+
+<p>
+  Any arrow keys work. 
+  Only one option can be selected. 
+  Filter by typing a letter. 
+  Headings are not selectable.
+</p>
+<pf-simple-list>
+  <pf-simple-list-group>
+    <span slot="group-heading">Breakfast</span>
+    <pf-simple-list-option>Eggs</pf-simple-list-option>
+    <pf-simple-list-option>Toast</pf-simple-list-option>
+    <pf-simple-list-option>Waffles</pf-simple-list-option>
+  </pf-simple-list-group>
+  <pf-simple-list-group>
+    <span slot="group-heading">Lunch</span>
+    <pf-simple-list-option>Salad</pf-simple-list-option>
+    <pf-simple-list-option>Sandwich</pf-simple-list-option>
+    <pf-simple-list-option>Soup</pf-simple-list-option>
+  </pf-simple-list-group>
+</pf-simple-list>
