+++ elements/pf-simple-list/demo/multiple-selections.html
@@ -1,0 +1,19 @@
+<link rel="stylesheet" href="demo.css">
+<script type="module" src="pf-simple-list.js"></script>
+
+<p>
+  Multiple options can be selected. Any arrow keys work. 
+  <kbd>Shift</kbd> will toggling off multiple items. 
+  <kbd>Ctrl+A</kbd> will toggle selection on all items. 
+  Filter by typing a letter.
+</p>
+<pf-simple-list multiple>
+  <pf-simple-list-option>Blue</pf-simple-list-option>
+  <pf-simple-list-option>Green</pf-simple-list-option>
+  <pf-simple-list-option>Magenta</pf-simple-list-option>
+  <pf-simple-list-option>Orange</pf-simple-list-option>
+  <pf-simple-list-option>Purple</pf-simple-list-option>
+  <pf-simple-list-option>Pink</pf-simple-list-option>
+  <pf-simple-list-option>Red</pf-simple-list-option>
+  <pf-simple-list-option>Yellow</pf-simple-list-option>
+</pf-simple-list>
