+++ elements/pf-simple-list/demo/orientation.html
@@ -1,0 +1,28 @@
+<link rel="stylesheet" href="demo.css">
+<script type="module" src="pf-simple-list.js"></script>
+
+<h2>orientation="vertical"</h2>
+<p>Only vertical arrow keys work. Filter by typing a letter.</p>
+<pf-simple-list orientation="vertical">
+  <pf-simple-list-option>Blue</pf-simple-list-option>
+  <pf-simple-list-option>Green</pf-simple-list-option>
+  <pf-simple-list-option>Magenta</pf-simple-list-option>
+  <pf-simple-list-option>Orange</pf-simple-list-option>
+  <pf-simple-list-option>Purple</pf-simple-list-option>
+  <pf-simple-list-option>Pink</pf-simple-list-option>
+  <pf-simple-list-option>Red</pf-simple-list-option>
+  <pf-simple-list-option>Yellow</pf-simple-list-option>
+</pf-simple-list>
+
+<h2>orientation="horizontal"</h2>
+<p>Only horizontal arrow keys work. Filter by typing a letter.</p>
+<pf-simple-list orientation="horizontal">
+  <pf-simple-list-option>Blue</pf-simple-list-option>
+  <pf-simple-list-option>Green</pf-simple-list-option>
+  <pf-simple-list-option>Magenta</pf-simple-list-option>
+  <pf-simple-list-option>Orange</pf-simple-list-option>
+  <pf-simple-list-option>Purple</pf-simple-list-option>
+  <pf-simple-list-option>Pink</pf-simple-list-option>
+  <pf-simple-list-option>Red</pf-simple-list-option>
+  <pf-simple-list-option>Yellow</pf-simple-list-option>
+</pf-simple-list>
